requirements.txt
numpy
scipy
biopython
MDAnalysis